pyproject.toml
[project]
name = "otstack"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "gitpython>=3.1.46",
    "pygithub>=2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a6",
]

[tool.ruff]
line-length = 88
src = ["src", "test"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["src"]
